+++ src/check_roms/test_check_by_name.rs
@@ -1,0 +1,73 @@
+use super::super::config::*;
+use super::super::import_dats;
+use super::super::import_roms;
+use super::*;
+use std::path::{Path, PathBuf};
+use tempfile::{NamedTempFile, TempDir};
+use tokio::fs;
+use tokio::io::AsyncWriteExt;
+
+#[tokio::test]
+async fn test() {
+    // given
+    let _guard = MUTEX.lock().await;
+
+    let test_directory = Path::new("tests");
+    let progress_bar = ProgressBar::hidden();
+
+    let db_file = NamedTempFile::new().unwrap();
+    let pool = establish_connection(db_file.path().to_str().unwrap()).await;
+    let mut connection = pool.acquire().await.unwrap();
+
+    let rom_directory = TempDir::new_in(test_directory).unwrap();
+    set_rom_directory(&mut connection, PathBuf::from(rom_directory.path())).await;
+    let tmp_directory = TempDir::new_in(test_directory).unwrap();
+    let tmp_directory =
+        set_tmp_directory(&mut connection, PathBuf::from(tmp_directory.path())).await;
+
+    let matches = import_dats::subcommand()
+        .get_matches_from(["import-dats", "tests/Test System (20200721).dat"]);
+    import_dats::main(&mut connection, &matches, &progress_bar)
+        .await
+        .unwrap();
+
+    let romfile_path = tmp_directory.join("Test Game (USA, Europe).rom");
+    fs::copy(
+        test_directory.join("Test Game (USA, Europe).rom"),
+        &romfile_path,
+    )
+    .await
+    .unwrap();
+
+    let system = find_systems(&mut connection).await.remove(0);
+
+    let matches = import_roms::subcommand()
+        .get_matches_from(["import-roms", romfile_path.as_os_str().to_str().unwrap()]);
+    import_roms::main(&mut connection, &matches, &progress_bar)
+        .await
+        .unwrap();
+
+    let romfile = find_romfiles(&mut connection).await.remove(0);
+    let mut file = fs::OpenOptions::new()
+        .write(true)
+        .open(rom_directory.path().join(&romfile.path))
+        .await
+        .unwrap();
+    file.write_all(b"00000000").await.unwrap();
+    file.sync_all().await.unwrap();
+
+    // when
+    let matches =
+        super::subcommand().get_matches_from(["check-roms", "--system", system.name.as_str()]);
+    super::main(&mut connection, &matches, &progress_bar)
+        .await
+        .unwrap();
+
+    // then
+    let mut romfiles = find_romfiles(&mut connection).await;
+    assert_eq!(romfiles.len(), 1);
+
+    let romfile = romfiles.remove(0);
+    assert!(romfile.path.contains("/Trash/"));
+    assert!(&rom_directory.path().join(&romfile.path).is_file());
+}